+++ src/bytecode.rs
@@ -162,6 +162,7 @@ pub enum Op {
     },
     Call(usize),
     CallSpread(Vec<CallArgumentKind>),
+    PipelineCall(Vec<CallArgumentKind>),
     TryMatch {
         pattern: MatchPattern,
         bindings: usize,